tekken pipeline for realtime

+++ owl_wms/models/tekken_rft_v2.py
@@ -51,7 +51,7 @@ class TekkenRFTCoreV2(nn.Module):
         # Generate action embeddings from button presses.
         action_tokens = self.action_embed(button_presses)  # [B, T, 8, D_model]
         action_emb = action_tokens.mean(dim=2)  # [B, T, D_model]
-        cond_emb = t_cond + action_emb  # [B, T, D_model]
+        cond_emb = (t_cond + action_emb).to(x.dtype)  # [B, T, D_model]
 
         # if not self.uncond and has_controls is not None:
             # Zero out embeddings where has_controls is False for CFG.
